+++ Silverware/src/drv_i2c.h
@@ -1,5 +1,4 @@
-
-
+#include <stdint.h>
 
 
 
@@ -7,12 +6,10 @@
 
 
 void i2c_init( void);
-int i2c_readdata( int reg, int *data, int size );
+int i2c_readdata( int reg, uint8_t *data, int size );
 int i2c_readreg( int reg );
 void i2c_writereg( int reg ,int data);
 			
-
-
